Improve name encoding
Now the sympy code output is a bit more readable
